refactor: Update test suite structure and remove obsolete tests
- Adjusted test discovery path in index.ts to include both suite and promptStore directories.
- Removed outdated test files for FileOperationManager and integration tests to streamline the codebase.
- Enhanced test organization by eliminating unnecessary files, improving maintainability.
