+++ wu-wei/src/test/suite/index.ts
@@ -10,11 +10,14 @@ export function run(): Promise<void> {
         timeout: 10000
     });
 
-    const testsRoot = path.resolve(__dirname, '.');
+    const testsRoot = path.resolve(__dirname, '..');
 
     return new Promise((c, e) => {
-        // Only look for tests in the suite directory
-        const testFiles = glob.sync('**/*.test.js', { cwd: testsRoot });
+        // Look for tests in suite and promptStore directories
+        const testFiles = glob.sync('**/*.test.js', {
+            cwd: testsRoot,
+            ignore: ['**/node_modules/**']
+        });
 
         // Add files to the test suite
         testFiles.forEach((f: string) => mocha.addFile(path.resolve(testsRoot, f)));
